requirements.txt
pandas
numpy
scikit-learn
matplotlib
plotly
jupyter
ipykernel
pyyaml
tqdm
rank-bm25
sentence-transformers
datasets
streamlit
openai
